refactor chat UI
